backend/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
defusedxml==0.7.1
Django==3.2.9
django-environ==0.8.1
django-extra-fields==3.0.2
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.8.0
gunicorn==20.0.4
djoser==2.1.0
flake8==4.0.1
idna==3.3
importlib-metadata==1.7.0
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.1.1
Pillow==8.4.0
psycopg2-binary==2.9.2
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sorl-thumbnail==12.7.0
sqlparse==0.4.2
typing_extensions==4.0.0
uritemplate==4.1.1
urllib3==1.26.7
zipp==3.6.0
